Explicitly set szl_api_version = 2.
There is still some Sawzall code running in production as a business logic execution engine in GMail and Gaia.  We need to manually finish the proto2 conversion in order to release a new Sawzall binary that is compatible with new changes to protocol buffers.

More information:  ​go/sawzall-proto2-lsc and go/sawzall-proto2

Tested:
    tap_presubmit: http://test/OCL:315701415:BASE:316210669:1592040780158:d61d033f
    Some tests failed; test failures are believed to be unrelated to this CL
PiperOrigin-RevId: 316462109